Remove the rebuild action from the derivatives lambda
